# iteh code/eventapi/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EventHub API",
        "description": "API dokumentacija za EventHub aplikaciju",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/event-participations": {
            "get": {
                "tags": [
                    "Event Participations"
                ],
                "summary": "Prikaz učestvovanja na događajima",
                "description": "ADMIN/ORGANIZATOR vide sva učestvovanja, ostali vide samo svoja.",
                "operationId": "e43d6defa40dfa4b073ec722217c82fb",
                "responses": {
                    "200": {
                        "description": "Uspešno vraćena lista učestvovanja"
                    },
                    "401": {
                        "description": "Neautorizovan pristup"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Event Participations"
                ],
                "summary": "Prijava korisnika na događaj",
                "operationId": "81e6d907ee00f4b0cc05338a53eb87d4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "idEvent",
                                    "status",
                                    "registeredAt"
                                ],
                                "properties": {
                                    "idEvent": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "REGISTERED"
                                    },
                                    "registeredAt": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2026-04-10 12:00:00"
                                    },
                                    "cancelledAt": {
                                        "type": "string",
                                        "format": "date-time",
                                        "nullable": true
                                    },
                                    "attendanceMarkedAt": {
                                        "type": "string",
                                        "format": "date-time",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Učestvovanje uspešno kreirano"
                    },
                    "401": {
                        "description": "Neautorizovan pristup"
                    },
                    "422": {
                        "description": "Greška validacije ili korisnik već prijavljen na događaj"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/event-participations/{id}": {
            "get": {
                "tags": [
                    "Event Participations"
                ],
                "summary": "Prikaz jednog učestvovanja na događaju",
                "operationId": "6fae780a579a77c797876a600218d322",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID učestvovanja (idParticipation)",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Uspešno vraćeno učestvovanje"
                    },
                    "401": {
                        "description": "Neautorizovan pristup"
                    },
                    "403": {
                        "description": "Zabranjen pristup"
                    },
                    "404": {
                        "description": "Učestvovanje nije pronađeno"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Event Participations"
                ],
                "summary": "Izmena učestvovanja na događaju",
                "operationId": "06001561716a71460b96311a5a3decc5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID učestvovanja (idParticipation)",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "idEvent": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "CANCELLED"
                                    },
                                    "registeredAt": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2026-04-10 12:00:00"
                                    },
                                    "cancelledAt": {
                                        "type": "string",
                                        "format": "date-time",
                                        "nullable": true
                                    },
                                    "attendanceMarkedAt": {
                                        "type": "string",
                                        "format": "date-time",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Učestvovanje uspešno izmenjeno"
                    },
                    "401": {
                        "description": "Neautorizovan pristup"
                    },
                    "403": {
                        "description": "Zabranjen pristup"
                    },
                    "404": {
                        "description": "Učestvovanje nije pronađeno"
                    },
                    "422": {
                        "description": "Greška validacije"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Event Participations"
                ],
                "summary": "Brisanje učestvovanja na događaju",
                "operationId": "b22c550d09e28c1895cf23378004e28b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID učestvovanja (idParticipation)",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Učestvovanje uspešno obrisano"
                    },
                    "401": {
                        "description": "Neautorizovan pristup"
                    },
                    "403": {
                        "description": "Zabranjen pristup"
                    },
                    "404": {
                        "description": "Učestvovanje nije pronađeno"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}